% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cols}
\alias{check_cols}
\title{Check columns of two data frames match
Easily check to make sure data frames have the same columns, no matter the
order. \code{check_cols} returns nothing unless there is a mismatch, in which it
throws an error. It does not check attributes or class.}
\usage{
check_cols(df1, df2)
}
\arguments{
\item{df1}{First data frame}

\item{df2}{Second data frame}
}
\value{
Throws error if columns don't match
}
\description{
Check columns of two data frames match
Easily check to make sure data frames have the same columns, no matter the
order. \code{check_cols} returns nothing unless there is a mismatch, in which it
throws an error. It does not check attributes or class.
}
\examples{
\dontrun{
check_cols(df1, df2)}
}
