% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compares.R
\name{check_equal}
\alias{check_equal}
\title{Compare two data frames}
\usage{
check_equal(df, df_new, df_names, ...)
}
\arguments{
\item{df}{Data frame source or "old"}

\item{df_new}{New data frame}

\item{df_names}{Vector of names to rename df to match df_new}
}
\value{
TRUE if they match, otherwise tibble of mismatched rows
}
\description{
Compare two data frames
}
\examples{
\dontrun{
check_equal(df |> dplyr::arrange(id), df_source |> dplyr::arrnage(id), df_names = names(df))
}
}
