% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_unique_key}
\alias{check_unique_key}
\title{Check if key is unique}
\usage{
check_unique_key(id_col)
}
\arguments{
\item{id_col}{Column believed to have unique values similar to a primary key}
}
\value{
Throws error if id_col is not unique
}
\description{
Check if key is unique
}
\examples{
check_unique_key(c(1, 2, 3))
}
