% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{deleteData}
\alias{deleteData}
\title{Delete from DB}
\usage{
deleteData(conn, table, var, condition, operation = "=")
}
\arguments{
\item{conn}{Connection}

\item{table}{Fully qualified table name}

\item{var}{Variable used to identify row}

\item{condition}{Condition to delete}

\item{operation}{Operation

This function has been designed to be more restrictive since deleting is less common
and more destructive. If you want to delete multiple rows, paste lookup values in the condition
attribute or use functional}
}
\value{
The number of rows affected
}
\description{
Delete from DB
}
\examples{
\dontrun{
deleteData(conn, "data.tbl_metrics, "id", "234")
deleteData(conn, "data.tbl_metrics, "id", paste("(", paste(df$id, collapse = ","), ")"))
}
}
