% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{newData}
\alias{newData}
\title{Write New Data to Database}
\usage{
newData(conn, schema, table, vals)
}
\arguments{
\item{conn}{Database connection}

\item{schema}{Schema of the table writing to}

\item{table}{Table to write to}

\item{vals}{Val or data frame that will be written to the table}
}
\value{
The number of rows affected
}
\description{
Write New Data to Database
}
\examples{
\dontrun{
newData(conn, data, "tbl_metrics", df)
}
}
