% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quote_sql}
\alias{quote_sql}
\title{Replace single quote with two single quotes}
\usage{
quote_sql(text)
}
\arguments{
\item{text}{Text value that needs to be cleaned}
}
\value{
Character value with SQL-escaped '
}
\description{
Replace single quote with two single quotes
}
