% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{updateData}
\alias{updateData}
\title{Update Data in Database}
\usage{
updateData(conn, table, df, var, condition, operation = "=")
}
\arguments{
\item{conn}{Conection to Db}

\item{table}{Fully qualified table name to write to}

\item{df}{Dataframe or tibble of data to update}

\item{var}{Variable to condition on}

\item{condition}{Condition}

\item{operation}{Operation for condition}
}
\value{
The number of rows affected
}
\description{
Update Data in Database
}
\examples{
\donotrun{
updateData(conn, data.tbl_metrics, df, id, 261, "=")
}
}
