% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compares.R
\name{check_equal}
\alias{check_equal}
\title{Compare two data frames}
\usage{
check_equal(df, df_source, df_names)
}
\arguments{
\item{df}{Data frame to compare to}

\item{df_source}{Data frame source}

\item{df_names}{Vector of names to rename df_source to match df

Note: You can use dplyr functions in the arguments to make sure the data frames
will match correctly. This is especially important when data is pulled from a db
since order is not guaranteed. This function is a wrapper for testthat::expect_equal
which will check class and values}
}
\value{
pass/fail from testthat::test_that
}
\description{
Compare two data frames
}
\examples{
\dontrun{
check_equal(df |> dplyr::arrange(id), df_source |> dplyr::arrnage(id), df_names = names(df))
}
}
