% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compares.R
\name{compares}
\alias{compares}
\alias{new_vals}
\alias{changed_vals}
\alias{old_vals}
\title{Changes between two data frames}
\usage{
new_vals(df_old, df_new, key)

changed_vals(df_old, df_new, key)

old_vals(df_old, df_new, key)
}
\arguments{
\item{df_old}{Current data frame}

\item{df_new}{New date frame that is being checked for new values}

\item{key}{Name of column containing unique values similar to primary key}
}
\value{
Data frame of values
}
\description{
Compares two data frames \code{df_old} and \code{df_new}.These functions were originally
intended for syncing a remote data source with a local one, but generally
work for comparing any data frames.
\itemize{
\item \code{new_vals(df_old, df_new, key)} finds all of the rows with a value in the
key column in \code{df_new} that is not in \code{df_old}
\item \code{changed_vals(df_old, df_new, key)} finds all of the rows that have the same
value in the key column but different data in the other columns
\item \code{old_vals(df_old, df_new, key)} finds all of the rows with a value in the
key column in \code{df_old} that is not in \code{df_new}
}
}
\examples{
\dontrun{
new_rows <- new_vals(df_old, df_new, "pk_col_name")
updated_rows <- changed_vals(df_old, df_new, "pk_col_name")
deleted_rows <- old_vals(df_old, df_new, "pk_col_name")
}
}
