% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_table}
\alias{get_table}
\title{Get Table/View of data}
\usage{
get_table(conn, tbl, vars, names = FALSE)
}
\arguments{
\item{conn}{DB connection}

\item{tbl}{table name}

\item{vars}{column names}

\item{names}{vector of column names if they need to be replaced}
}
\value{
tibble of database table
}
\description{
Get Table/View of data
}
\examples{
\dontrun{
get_table(conn, tbl_data, vars_vec)
}
}
